feat: add india to country_map visualization (#182)
* feat: add india to country_map visualization

India, through a merge conflict of some sort had been left out in the
main repo when country_map was migrated to this repo.

As part of the 0.34.0rc1 release process, this file got flagged as a
potential license issue. Moving it out of the source release and here is
the right thing to do.

* docs: enable country map

* fix: import india

* build: disable geojson parsing

* fix: import

* build: disable country map storybook